FIX - Reset Translations only when backend serves in standard language

+++ lib/ez_sql_mysqli.php
@@ -260,40 +260,48 @@ class ezSQL_mysqli extends ezSQLcore
 		//translations... only if update and not deepl Plugin
 		if($this->deepl == false)
 		{
-			if((stristr($query,"UPDATE") || stristr($query,"DELETE")) )
+			//checken if backend lang = standard lang - else go on - change nothing...
+			if($_SESSION['langdata_front']['lang_short'] == $_SESSION['dbp']['papoo_daten2']['0']['lang_frontend'])
 			{
-				if(empty($this->transDbs))
+				//print_r("");
+				if((stristr($query,"UPDATE") || stristr($query,"DELETE")) )
 				{
-					//check if translation relevant
-					$transsql = sprintf("SELECT * FROM %s",DB_PRAEFIX."trans_tabnames");
-					$this->transDbs = $this->get_results($transsql,ARRAY_A);
-				}
-
-				if(!empty($this->transDbs))
-				{
-					foreach($this->transDbs as $dbNames)
+					if(empty($this->transDbs))
 					{
-						if(stristr($query,$dbNames['trans_name_tab_name']))
-						{
-							//Oha - it is language and translation relevant...
-							$searchQuery = strtolower($query);
-							$sqa1 = explode("where",$searchQuery);
-							$sqa2 = explode("and",$sqa1['1']);
-							foreach ($sqa2 as $checkIds)
-							{
-								if(stristr($checkIds,$dbNames['trans_name_id_name']))
-								{
-									//finally the id is found
-									$id = (trim(preg_replace("/[^0-9]/", "", $checkIds)));
+						//check if translation relevant
+						$transsql = sprintf("SELECT * FROM %s",DB_PRAEFIX."trans_tabnames");
+						$this->transDbs = $this->get_results($transsql,ARRAY_A);
+					}
 
-									if(is_numeric($id) && $id > 0)
+					if(!empty($this->transDbs))
+					{
+						foreach($this->transDbs as $dbNames)
+						{
+							if(stristr($query,$dbNames['trans_name_tab_name']))
+							{
+								//Oha - it is language and translation relevant...
+								$searchQuery = strtolower($query);
+
+								//lets find the ids...
+								$sqa1 = explode("where",$searchQuery);
+								$sqa2 = explode("and",$sqa1['1']);
+								foreach ($sqa2 as $checkIds)
+								{
+									if(stristr($checkIds,$dbNames['trans_name_id_name']))
 									{
-										//now we can delete this entry from the translation Table - so it can be translated again..
-										$delSsql = sprintf("DELETE FROM %s WHERE trans_tab_name='%s' AND trans_id_id='%d'",
-											DB_PRAEFIX."trans_ids",
-											$dbNames['trans_name_tab_name'],
-											$id);
-										$this->query($delSsql);
+										//finally the id is found
+										$id = (trim(preg_replace("/[^0-9]/", "", $checkIds)));
+
+										//Should be numeric :-)
+										if(is_numeric($id) && $id > 0)
+										{
+											//now we can delete this entry from the translation Table - so it can be translated again..
+											$delSsql = sprintf("DELETE FROM %s WHERE trans_tab_name='%s' AND trans_id_id='%d'",
+												DB_PRAEFIX."trans_ids",
+												$dbNames['trans_name_tab_name'],
+												$id);
+											$this->query($delSsql);
+										}
 									}
 								}
 							}
